% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADAP.R
\name{ADAP.estimate}
\alias{ADAP.estimate}
\title{ADAP surrogate estimation}
\usage{
ADAP.estimate(ipdata,control,config)
}
\arguments{
\item{ipdata}{local data in data frame}

\item{control}{PDA control}

\item{config}{cloud configuration}
}
\value{
list(btilde = sol$par, Htilde = sol$hessian, site=control$mysite, site_size=nrow(ipdata))
}
\description{
ADAP surrogate estimation
}
\details{
step-3: construct and solve surrogate objective function at the master/lead site
}
\keyword{internal}
