% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADAP.R
\name{ADAP.initialize}
\alias{ADAP.initialize}
\title{ADAP initialize}
\usage{
ADAP.initialize(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
init
}
\description{
ADAP initialize
}
\keyword{internal}
