% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADAP_data}
\alias{ADAP_data}
\title{ADAP simulated data}
\format{
A list containing the following elements:
\describe{
  \item{sites}{site id, 300 'site1', 300 'site2', 300 'site3'}
  \item{status}{binary outcome of length 900}
  \item{x}{900 by 49 matrix generated by standard normal distribution, representing the covariates} 
}
}
\usage{
ADAP_data
}
\description{
A simulated data set for ADAP demonstration
}
\keyword{datasets}
