\name{branch}
\alias{branch}
\alias{burst}
\title{Branch Plot for Nested Effects}
\usage{branch(root, limb, pos=1:2, flip=FALSE)}
\arguments{
    \item{root}{root value}
    \item{limb}{limb value}
    \item{pos}{positions for root and limb}
    \item{flip}{flip to vertical display if TRUE}}
\value{Nothing returned.}
\description{Draws lines from root}
\examples{
#
\dontrun{
branch(root, limb)
}
}
\keyword{hplot}
% Converted by  version .
