% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLA.R
\name{COLA.estimate}
\alias{COLA.estimate}
\title{PDA COLA estimation}
\usage{
COLA.estimate(ipdata=NULL,control,config)
}
\arguments{
\item{ipdata}{no need}

\item{control}{PDA control}

\item{config}{cloud configuration}
}
\value{
list(est, se)
}
\description{
PDA COLA estimation
}
\details{
COLA estimation: 
(1) COLA-GLM
(2) COLA-GLM-H
(3) COLA-GLMM
}
\keyword{internal}
