% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLA.R
\name{COLA.initialize}
\alias{COLA.initialize}
\title{COLA initialize}
\usage{
COLA.initialize(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
init
}
\description{
COLA initialize
}
\references{
Qiong Wu, et al. (2025) COLA-GLM: Collaborative One-shot and Lossless Algorithms of Generalized Linear Models for Decentralized Observational Healthcare Data. 
   npj Digital Medicine. \cr
   Bingyu Zhang, et al (2025) A Lossless One-shot Distributed Algorithm for Addressing Heterogeneity in Multi-Site Generalized Linear Models.
   Journal of the American Medical Informatics Association (under revision). \cr
   Jiayi Tong, et al. (2025) Unlocking Efficiency in Real-world Collaborative Studies: A Multi-site International Study with Collaborative One-shot Lossless Algorithm for Generalized Linear Mixed Model.
   npj Digital Medicine. \cr
}
\keyword{internal}
