% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COLA_covid}
\alias{COLA_covid}
\title{COLA simulated data}
\format{
A data frame with 1500 rows and 6 variables:
\describe{
\item{site}{site, 600 'site1', 500 'site2', 400 'site3'}
\item{age}{binary age}
\item{sex}{binary sex}
\item{medical_condition}{binary medical condition}
\item{status}{binary outcome, COVID-19 status. This is the binary outcome for COLA logistic regression}
\item{visits}{poisson outcome, number of visits. This is the count outcome for COLA Poisson regression}
}
}
\usage{
COLA_covid
}
\description{
A simulated COVID-19 data set for Collaborative One-shot and Lossless Algorithms of generalized linear models (COLA-GLM)
}
\keyword{datasets}
