% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistCox.R
\name{combine_data}
\alias{combine_data}
\title{Combine and reformate data from multiple sites}
\usage{
combine_data(local_data, remote_data, col_time, col_event, col_X)
}
\arguments{
\item{local_data}{Data frame containing time-to-event, event indicator and covariates in the local machine.}

\item{remote_data}{List of data frames, each df containing time-to-event, event indicator and covariates in a remote machine.}

\item{col_time}{Integer, which col of the df is the time-to-event. time-to-event need to be positive, can be tied and right-censored (e.g. event=1)}

\item{col_event}{Integer, which col of the df is the event indicator. event=1 if event happens, event=0 if censored}

\item{col_X}{Vector of integers, which cols are the covariates. If not specified (i.e. col_X=0), use all col except col_time and col_event}
}
\value{
List of local_data and all_data
}
\description{
combine data (in data frame) from multiple machines together to list of time-to-event, event indicator and covariates.
}
\examples{
# See DistCox()
}
\author{
Chongliang Luo, Rui Duan, Yong Chen
}
