\name{Count}
\alias{Count}
\title{Counting Data Examples}
\description{These are examples of counting data analysis.}
\references{Venables and Ripley (1999) \emph{Modern Applied Statistics
	with S-PLUS}, 3rd ed., ch. 7.}
\examples{
example( GrowthCount )
example( Budworm )
example( BrandX )
} 
% Converted by  version .
\keyword{datasets}
