% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs}
\alias{cs}
\title{CrabSatellites data}
\format{
A data frame containing 173 observations on 4 variables.
\describe{
  \item{site}{Simulated site id, 85 'site1' and 88 'site2'.}
  \item{satellites}{Number of satellites. Treated as (zero-inflated) count outcome in ODAH}
  \item{width}{Carapace width (cm).}
  \item{weight}{Weight (kg).} 
}
}
\source{
\url{https://rdrr.io/rforge/countreg/man/CrabSatellites.html}
}
\usage{
cs
}
\description{
A data set modified from the CrabSatellites data in countreg package (see demo(ODAH)).
}
\keyword{datasets}
