% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGEM.R
\name{dGEM.derive}
\alias{dGEM.derive}
\title{dGEM hospital-specific effect derivation}
\usage{
dGEM.derive(ipdata,control,config,hosdata)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}

\item{hosdata}{hospital-level data}
}
\value{
hospital_effect
}
\description{
dGEM hospital-specific effect derivation
}
\keyword{internal}
