% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGEM.R
\name{dGEM.estimate}
\alias{dGEM.estimate}
\title{dGEM standardized event rate estimation}
\usage{
dGEM.estimate(ipdata,control,config)
}
\arguments{
\item{ipdata}{local data in data frame}

\item{control}{PDA control}

\item{config}{cloud configuration}
}
\value{
event rate
}
\description{
dGEM standardized event rate estimation
}
\details{
step-3:
}
\keyword{internal}
