% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGEM.R
\name{dGEM.initialize}
\alias{dGEM.initialize}
\title{dGEM initialize}
\usage{
dGEM.initialize(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
init
}
\description{
dGEM initialize
}
\references{
NA
}
\keyword{internal}
