% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGEM.R
\name{dGEM.synthesize}
\alias{dGEM.synthesize}
\title{PDA dGEM synthesize}
\usage{
dGEM.synthesize(ipdata,control,config)
}
\arguments{
\item{ipdata}{local data in data frame}

\item{control}{pda control}

\item{config}{pda cloud configuration}
}
\value{
list(final_event_rate=final_event_rate)
}
\description{
PDA dGEM synthesize
}
\details{
Synthesis to get the standardized mortality rate
}
\keyword{internal}
