% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisC2o.R
\name{DisC2o.AIPWestimate}
\alias{DisC2o.AIPWestimate}
\title{DisC2o AIPW estimate of the ATE at each site}
\usage{
DisC2o.AIPWestimate(ipdata,control,config)
}
\arguments{
\item{ipdata}{local data in data frame}

\item{control}{pda control}

\item{config}{pda cloud configuration}
}
\value{
list(btilde=btilde,  Vtilde=Vtilde)
}
\description{
DisC2o AIPW estimate of the ATE at each site
}
\keyword{internal}
