% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisC2o.R
\name{DisC2o.OMderive}
\alias{DisC2o.OMderive}
\title{DisC2o_OM derivatives}
\usage{
DisC2o.OMderive(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
list(site=config$site_id, site_size = nrow(ipdata), logL_D1=logL_D1, logL_D2=logL_D2)
}
\description{
DisC2o_OM derivatives
}
\keyword{internal}
