% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisC2o.R
\name{DisC2o.OMestimate}
\alias{DisC2o.OMestimate}
\title{DisC2o outcome model surrogate estimation}
\usage{
DisC2o.OMestimate(ipdata,control,config)
}
\arguments{
\item{ipdata}{local data in data frame}

\item{control}{PDA control}

\item{config}{cloud configuration}
}
\value{
list(btilde = sol$par, Htilde = sol$hessian, site=control$mysite, site_size=nrow(ipdata))
}
\description{
DisC2o outcome model surrogate estimation
}
\keyword{internal}
