% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisC2o.R
\name{DisC2o.OMinitialize}
\alias{DisC2o.OMinitialize}
\title{DisC2o_OM initialize}
\usage{
DisC2o.OMinitialize(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
init
}
\description{
DisC2o_OM initialize
}
\keyword{internal}
