% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisC2o.R
\name{DisC2o.PSinitialize}
\alias{DisC2o.PSinitialize}
\title{DisC2o PS initialize}
\usage{
DisC2o.PSinitialize(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
init
}
\description{
DisC2o PS initialize
}
\references{
Tong J, et al. 2025. DisC2o-HD: Distributed causal inference with covariates shift for analyzing real-world high-dimensional data. Journal of Machine Learning Research. 2025;26(3):1-50.
}
\keyword{internal}
