% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisC2o.R
\name{DisC2o.synthesize}
\alias{DisC2o.synthesize}
\title{DisC2o AIPW estimate of the ATE, synthesizing all sites}
\usage{
DisC2o.synthesize(ipdata,control,config)
}
\arguments{
\item{ipdata}{local data in data frame}

\item{control}{pda control}

\item{config}{pda cloud configuration}
}
\value{
list(btilde=btilde,  Vtilde=Vtilde)
}
\description{
DisC2o AIPW estimate of the ATE, synthesizing all sites
}
\keyword{internal}
