% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLM.R
\name{DLM.estimate}
\alias{DLM.estimate}
\title{PDA DLM estimation}
\usage{
DLM.estimate(ipdata=NULL,control,config)
}
\arguments{
\item{ipdata}{no need}

\item{control}{PDA control}

\item{config}{cloud configuration}
}
\value{
list(bhat, sebhat, sigmahat, uhat, seuhat)
}
\description{
PDA DLM estimation
}
\details{
DLM estimation: 
(1) Linear model, 
(2) Linear model with fixed effects, 
(3) Linear model with random effects (Linear mixed model)
}
\keyword{internal}
