% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLM.R
\name{DLM.initialize}
\alias{DLM.initialize}
\title{DLM initialize}
\usage{
DLM.initialize(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
init
}
\description{
DLM initialize
}
\references{
Yixin Chen, et al. (2006) Regression cubes with lossless compression and aggregation. 
   IEEE Transactions on Knowledge and Data Engineering, 18(12), pp.1585-1599. \cr
   (DLMM) Chongliang Luo, et al. (2020) Lossless Distributed Linear Mixed Model with Application to Integration of Heterogeneous Healthcare Data.  
   medRxiv, \doi{10.1101/2020.11.16.20230730}. \cr
}
\keyword{internal}
