% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPQL.R
\name{DPQL.derive}
\alias{DPQL.derive}
\title{DPQL derive}
\usage{
DPQL.derive(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
list(SiX, SiXY, SiY, ni)
}
\description{
DPQL derive
}
\details{
This step calculated the intermediate aggregated data (XtWX, XtWY, and YtWY) for each site. 
          May need to be iterated several times until prespecified rounds are met.
}
\references{
Chongliang Luo, et al. (2021) dPQL: a lossless distributed algorithm for generalized linear mixed model 
                  with application to privacy-preserving hospital profiling. medRxiv, \doi{10.1101/2021.05.03.21256561}. \cr
           Chongliang Luo, et al. (2020) Lossless Distributed Linear Mixed Model with Application to Integration of Heterogeneous Healthcare Data.  
                 medRxiv, \doi{10.1101/2020.11.16.20230730}. \cr
}
\keyword{internal}
