% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPQL.R
\name{DPQL.estimate}
\alias{DPQL.estimate}
\title{PDA DPQL estimation}
\usage{
DPQL.estimate(ipdata=NULL,control,config)
}
\arguments{
\item{ipdata}{no need}

\item{control}{PDA control}

\item{config}{cloud configuration}
}
\value{
list(risk_factor, risk_factor_heterogeneity, bhat, sebhat, uhat, seuhat, Vhat)
}
\description{
PDA DPQL estimation
}
\details{
DPQL estimation: (iterative) weighted DLMM using AD from all sites
}
\references{
Chongliang Luo, et al. (2021) dPQL: a lossless distributed algorithm for generalized linear mixed model 
                  with application to privacy-preserving hospital profiling. medRxiv, \doi{10.1101/2021.05.03.21256561}. \cr
           Chongliang Luo, et al. (2020) Lossless Distributed Linear Mixed Model with Application to Integration of Heterogeneous Healthcare Data.  
                medRxiv, \doi{10.1101/2020.11.16.20230730}. \cr
}
\keyword{internal}
