% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmm.R
\name{estimatePool}
\alias{estimatePool}
\title{Pooled estimation for COLA-GLM and COLA-GLM-H}
\usage{
estimatePool(KSiteIPD, formula, family = "binomial", outcome_name, heter_intercept = FALSE)
}
\arguments{
\item{KSiteIPD}{\code{glm} data}

\item{formula}{\code{glm} formula}

\item{family}{\code{glm} family}

\item{outcome_name}{outcome name}

\item{heter_intercept}{Logical; if \code{TRUE}, includes site-specific intercepts to model heterogeneity across sites.}
}
\value{
A data frame with point estimates and standard errors for each coefficient
}
\description{
Performs pooled estimation for Generalized Linear Models (GLMs) using either the standard COLA-GLM approach 
or the heterogeneous intercept extension (COLA-GLM-H). This function supports binary and Poisson outcomes.
}
\keyword{internal}
