% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.R
\name{getCloudConfig}
\alias{getCloudConfig}
\title{gather cloud settings into one list}
\usage{
getCloudConfig()
}
\arguments{
\item{name}{of file}
}
\description{
gather cloud settings into one list
}
\author{
Chongliang Luo, Steven Vitale
}
