% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LATTE.R
\name{LATTE.estimate}
\alias{LATTE.estimate}
\title{LATTE LATTE.estimate}
\usage{
LATTE.estimate(init_data, control, config)
}
\arguments{
\item{init_data}{initialization data from LATTE.initialize}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
analysis results
}
\description{
LATTE conditional log-likelihood reconstruction at Lead Site
}
\keyword{internal}
