% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LATTE.R
\name{LATTE.initialize}
\alias{LATTE.initialize}
\title{LATTE initialize}
\usage{
LATTE.initialize(ipdata, control, config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
init object containing prepared data and PS model
}
\description{
LATTE (Lossless Aggregation for Treatment effect estimation) initialization: Propensity Score stratification/matching at Lead site
}
\keyword{internal}
