% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LATTE_ADRD}
\alias{LATTE_ADRD}
\title{LATTE simulated data}
\format{
A data frame with 1224 rows and 326 variables:
\describe{
  \item{ID}{Unique patient identifier}
  \item{Stroke...Transient.Ischemic.Attack}{History of Stroke or Transient Ischemic Attack, binary indicator}
  \item{Acquired.Hypothyroidism}{History of Acquired Hypothyroidism, binary indicator}
  \item{Fibromyalgia..Chronic.Pain.and.Fatigue}{History of Fibromyalgia, Chronic Pain and Fatigue, binary indicator}
  \item{RA.OA..Rheumatoid.Arthritis..Osteoarthritis.}{History of RA/OA (Rheumatoid Arthritis / Osteoarthritis), binary indicator}
  \item{Hypertension}{History of Hypertension, binary indicator}
  \item{Anxiety.Disorders}{History of Anxiety Disorders, binary indicator}
  \item{Chronic.Obstructive.Pulmonary.Disease.and.Bronchiectasis}{History of Chronic Obstructive Pulmonary Disease (COPD) and Bronchiectasis, binary indicator}
  \item{Asthma}{History of Asthma, binary indicator}
  \item{Traumatic.Brain.Injury.and.Nonpsychotic.Mental.Disorders.due.to.Brain.Damage}{History of Traumatic Brain Injury and Nonpsychotic Mental Disorders due to Brain Damage, binary indicator}
  \item{Sleep.disorders}{History of Sleep disorders, binary indicator}
  \item{ADHD..Conduct.Disorders..and.Hyperkinetic.Syndrome}{History of ADHD, Conduct Disorders, and Hyperkinetic Syndrome, binary indicator}
  \item{Cataract}{History of Cataract, binary indicator}
  \item{Migraine.and.Chronic.Headache}{History of Migraine and Chronic Headache, binary indicator}
  \item{Depressive.Disorders}{History of Depressive Disorders, binary indicator}
  \item{Hyperlipidemia}{History of Hyperlipidemia, binary indicator}
  \item{Sensory...Deafness.and.Hearing.Impairment}{History of Sensory Deafness and Hearing Impairment, binary indicator}
  \item{Female...Male.Breast.Cancer}{History of Female/Male Breast Cancer, binary indicator}
  \item{Personality.Disorders}{History of Personality Disorders, binary indicator}
  \item{Anemia}{History of Anemia, binary indicator}
  \item{Chronic.Kidney.Disease}{History of Chronic Kidney Disease, binary indicator}
  \item{Schizophrenia.and.Other.Psychotic.Disorders}{History of Schizophrenia and Other Psychotic Disorders, binary indicator}
  \item{Glaucoma}{History of Glaucoma, binary indicator}
  \item{Peripheral.Vascular.Disease..PVD.}{History of Peripheral Vascular Disease (PVD), binary indicator}
  \item{Heart.Failure}{History of Heart Failure, binary indicator}
  \item{Pressure.and.Chronic.Ulcers}{History of Pressure and Chronic Ulcers, binary indicator}
  \item{Obesity}{History of Obesity, binary indicator}
  \item{Diabetes}{History of Diabetes, binary indicator}
  \item{Mobility.Impairments}{History of Mobility Impairments, binary indicator}
  \item{Benign.Prostatic.Hyperplasia}{History of Benign Prostatic Hyperplasia, binary indicator}
  \item{Drug.Use.Disorders}{History of Drug Use Disorders, binary indicator}
  \item{Alcohol.Use.Disorders}{History of Alcohol Use Disorders, binary indicator}
  \item{Post.Traumatic.Stress.Disorder..PTSD.}{History of Post Traumatic Stress Disorder (PTSD), binary indicator}
  \item{Atrial.Fibrillation}{History of Atrial Fibrillation, binary indicator}
  \item{Tobacco.Use}{History of Tobacco Use, binary indicator}
  \item{Ischemic.Heart.Disease}{History of Ischemic Heart Disease, binary indicator}
  \item{Liver.Disease..Cirrhosis.and.Other.Liver.Conditions..except.Viral.Hepatitis.}{History of Liver Disease (Cirrhosis and Other Liver Conditions except Viral Hepatitis), binary indicator}
  \item{Sensory...Blindness.and.Visual.Impairment}{History of Sensory Blindness and Visual Impairment, binary indicator}
  \item{Bipolar.Disorder}{History of Bipolar Disorder, binary indicator}
  \item{Depression}{History of Depression, binary indicator}
  \item{Prostate.Cancer}{History of Prostate Cancer, binary indicator}
  \item{Acute.Myocardial.Infarction}{History of Acute Myocardial Infarction, binary indicator}
  \item{Hip.Pelvic.Fracture}{History of Hip/Pelvic Fracture, binary indicator}
  \item{Other.Developmental.Delays}{History of Other Developmental Delays, binary indicator}
  \item{Viral.Hepatitis..General.}{History of Viral Hepatitis (General), binary indicator}
  \item{Sickle.Cell.Disease}{History of Sickle Cell Disease, binary indicator}
  \item{Multiple.Sclerosis.and.Transverse.Myelitis}{History of Multiple Sclerosis and Transverse Myelitis, binary indicator}
  \item{Leukemias.and.Lymphomas}{History of Leukemias and Lymphomas, binary indicator}
  \item{Opioid.Use.Disorder}{History of Opioid Use Disorder, binary indicator}
  \item{Colorectal.Cancer}{History of Colorectal Cancer, binary indicator}
  \item{Epilepsy}{History of Epilepsy, binary indicator}
  \item{Osteoporosis}{History of Osteoporosis, binary indicator}
  \item{Intellectual.Disabilities.and.Related.Conditions}{History of Intellectual Disabilities and Related Conditions, binary indicator}
  \item{Spinal.Cord.Injury}{History of Spinal Cord Injury, binary indicator}
  \item{Endometrial.Cancer}{History of Endometrial Cancer, binary indicator}
  \item{Spina.Bifida.and.Other.Congenital.Anomalies.of.the.Nervous.System}{History of Spina Bifida and Other Congenital Anomalies of the Nervous System, binary indicator}
  \item{Learning.Disabilities}{History of Learning Disabilities, binary indicator}
  \item{Periodontitis}{History of Periodontitis, binary indicator}
  \item{Lung.Cancer}{History of Lung Cancer, binary indicator}
  \item{Cystic.Fibrosis.and.Other.Metabolic.Developmental.Disorders}{History of Cystic Fibrosis and Other Metabolic Developmental Disorders, binary indicator}
  \item{Cerebral.Palsy}{History of Cerebral Palsy, binary indicator}
  \item{Human.Immunodeficiency.Virus.and.or.Acquired.Immunodeficiency.Syndrome..HIV.AIDS.}{History of Human Immunodeficiency Virus and/or Acquired Immunodeficiency Syndrome (HIV/AIDS), binary indicator}
  \item{Menopause}{History of Menopause, binary indicator}
  \item{Muscular.Dystrophy}{History of Muscular Dystrophy, binary indicator}
  \item{Autism.Spectrum.Disorders}{History of Autism Spectrum Disorders, binary indicator}
  \item{atorvastatin}{History of atorvastatin use, binary indicator}
  \item{hydrochlorothiazide}{History of hydrochlorothiazide use, binary indicator}
  \item{amlodipine}{History of amlodipine use, binary indicator}
  \item{aspirin}{History of aspirin use, binary indicator}
  \item{metoprolol}{History of metoprolol use, binary indicator}
  \item{levothyroxine}{History of levothyroxine use, binary indicator}
  \item{metformin}{History of metformin use, binary indicator}
  \item{lisinopril}{History of lisinopril use, binary indicator}
  \item{simvastatin}{History of simvastatin use, binary indicator}
  \item{sodium.chloride}{History of sodium chloride use, binary indicator}
  \item{omeprazole}{History of omeprazole use, binary indicator}
  \item{albuterol}{History of albuterol use, binary indicator}
  \item{potassium.chloride}{History of potassium chloride use, binary indicator}
  \item{sertraline}{History of sertraline use, binary indicator}
  \item{lidocaine}{History of lidocaine use, binary indicator}
  \item{furosemide}{History of furosemide use, binary indicator}
  \item{losartan}{History of losartan use, binary indicator}
  \item{donepezil}{History of donepezil use, binary indicator}
  \item{cholecalciferol}{History of cholecalciferol use, binary indicator}
  \item{ondansetron}{History of ondansetron use, binary indicator}
  \item{oxycodone}{History of oxycodone use, binary indicator}
  \item{lorazepam}{History of lorazepam use, binary indicator}
  \item{glucose.oxidase}{History of glucose oxidase use, binary indicator}
  \item{prednisone}{History of prednisone use, binary indicator}
  \item{fluticasone}{History of fluticasone use, binary indicator}
  \item{gabapentin}{History of gabapentin use, binary indicator}
  \item{rosuvastatin}{History of rosuvastatin use, binary indicator}
  \item{tamsulosin}{History of tamsulosin use, binary indicator}
  \item{fentanyl}{History of fentanyl use, binary indicator}
  \item{carbidopa}{History of carbidopa use, binary indicator}
  \item{pantoprazole}{History of pantoprazole use, binary indicator}
  \item{escitalopram}{History of escitalopram use, binary indicator}
  \item{insulin.aspart..human}{History of insulin aspart (human) use, binary indicator}
  \item{clonazepam}{History of clonazepam use, binary indicator}
  \item{carvedilol}{History of carvedilol use, binary indicator}
  \item{heparin}{History of heparin use, binary indicator}
  \item{bupropion}{History of bupropion use, binary indicator}
  \item{polyethylene.glycol.3350}{History of polyethylene glycol 3350 use, binary indicator}
  \item{pravastatin}{History of pravastatin use, binary indicator}
  \item{docusate}{History of docusate use, binary indicator}
  \item{vitamin.B12}{History of vitamin B12 use, binary indicator}
  \item{clopidogrel}{History of clopidogrel use, binary indicator}
  \item{atenolol}{History of atenolol use, binary indicator}
  \item{glipizide}{History of glipizide use, binary indicator}
  \item{ascorbic.acid}{History of ascorbic acid use, binary indicator}
  \item{glucose}{History of glucose use, binary indicator}
  \item{insulin.glargine}{History of insulin glargine use, binary indicator}
  \item{zolpidem}{History of zolpidem use, binary indicator}
  \item{warfarin}{History of warfarin use, binary indicator}
  \item{azithromycin}{History of azithromycin use, binary indicator}
  \item{trazodone}{History of trazodone use, binary indicator}
  \item{esomeprazole}{History of esomeprazole use, binary indicator}
  \item{sennosides..USP}{History of sennosides (USP) use, binary indicator}
  \item{propofol}{History of propofol use, binary indicator}
  \item{iopamidol}{History of iopamidol use, binary indicator}
  \item{diltiazem}{History of diltiazem use, binary indicator}
  \item{ibuprofen}{History of ibuprofen use, binary indicator}
  \item{dexamethasone}{History of dexamethasone use, binary indicator}
  \item{tramadol}{History of tramadol use, binary indicator}
  \item{amoxicillin}{History of amoxicillin use, binary indicator}
  \item{midazolam}{History of midazolam use, binary indicator}
  \item{lansoprazole}{History of lansoprazole use, binary indicator}
  \item{citalopram}{History of citalopram use, binary indicator}
  \item{valsartan}{History of valsartan use, binary indicator}
  \item{ciprofloxacin}{History of ciprofloxacin use, binary indicator}
  \item{famotidine}{History of famotidine use, binary indicator}
  \item{calcium.carbonate}{History of calcium carbonate use, binary indicator}
  \item{finasteride}{History of finasteride use, binary indicator}
  \item{levofloxacin}{History of levofloxacin use, binary indicator}
  \item{sulfamethoxazole}{History of sulfamethoxazole use, binary indicator}
  \item{duloxetine}{History of duloxetine use, binary indicator}
  \item{alprazolam}{History of alprazolam use, binary indicator}
  \item{naproxen}{History of naproxen use, binary indicator}
  \item{levetiracetam}{History of levetiracetam use, binary indicator}
  \item{ranitidine}{History of ranitidine use, binary indicator}
  \item{triamcinolone}{History of triamcinolone use, binary indicator}
  \item{salmeterol}{History of salmeterol use, binary indicator}
  \item{folic.acid}{History of folic acid use, binary indicator}
  \item{nifedipine}{History of nifedipine use, binary indicator}
  \item{ferrous.sulfate}{History of ferrous sulfate use, binary indicator}
  \item{morphine}{History of morphine use, binary indicator}
  \item{hydralazine}{History of hydralazine use, binary indicator}
  \item{montelukast}{History of montelukast use, binary indicator}
  \item{magnesium.sulfate}{History of magnesium sulfate use, binary indicator}
  \item{methylphenidate}{History of methylphenidate use, binary indicator}
  \item{hydrocortisone}{History of hydrocortisone use, binary indicator}
  \item{latanoprost}{History of latanoprost use, binary indicator}
  \item{quetiapine}{History of quetiapine use, binary indicator}
  \item{metronidazole}{History of metronidazole use, binary indicator}
  \item{diphenhydramine}{History of diphenhydramine use, binary indicator}
  \item{memantine}{History of memantine use, binary indicator}
  \item{methylprednisolone}{History of methylprednisolone use, binary indicator}
  \item{doxycycline}{History of doxycycline use, binary indicator}
  \item{fluoxetine}{History of fluoxetine use, binary indicator}
  \item{paroxetine}{History of paroxetine use, binary indicator}
  \item{gadobenate}{History of gadobenate use, binary indicator}
  \item{propranolol}{History of propranolol use, binary indicator}
  \item{ramipril}{History of ramipril use, binary indicator}
  \item{ezetimibe}{History of ezetimibe use, binary indicator}
  \item{allopurinol}{History of allopurinol use, binary indicator}
  \item{enoxaparin}{History of enoxaparin use, binary indicator}
  \item{apixaban}{Binary indicator for history of apixaban}
  \item{sildenafil}{Binary indicator for history of sildenafil}
  \item{oxybutynin}{Binary indicator for history of oxybutynin}
  \item{insulin.lispro}{Binary indicator for history of insulin lispro}
  \item{melatonin}{Binary indicator for history of melatonin}
  \item{tadalafil}{Binary indicator for history of tadalafil}
  \item{hydromorphone}{Binary indicator for history of hydromorphone}
  \item{ipratropium}{Binary indicator for history of ipratropium}
  \item{cephalexin}{Binary indicator for history of cephalexin}
  \item{lovastatin}{Binary indicator for history of lovastatin}
  \item{mirtazapine}{Binary indicator for history of mirtazapine}
  \item{venlafaxine}{Binary indicator for history of venlafaxine}
  \item{fenofibrate}{Binary indicator for history of fenofibrate}
  \item{guaifenesin}{Binary indicator for history of guaifenesin}
  \item{estradiol}{Binary indicator for history of estradiol}
  \item{nitroglycerin}{Binary indicator for history of nitroglycerin}
  \item{pregabalin}{Binary indicator for history of pregabalin}
  \item{lamotrigine}{Binary indicator for history of lamotrigine}
  \item{enalapril}{Binary indicator for history of enalapril}
  \item{ergocalciferol}{Binary indicator for history of ergocalciferol}
  \item{ketoconazole}{Binary indicator for history of ketoconazole}
  \item{spironolactone}{Binary indicator for history of spironolactone}
  \item{cyclobenzaprine}{Binary indicator for history of cyclobenzaprine}
  \item{meloxicam}{Binary indicator for history of meloxicam}
  \item{cetirizine}{Binary indicator for history of cetirizine}
  \item{alendronate}{Binary indicator for history of alendronate}
  \item{nortriptyline}{Binary indicator for history of nortriptyline}
  \item{bisacodyl}{Binary indicator for history of bisacodyl}
  \item{sitagliptin}{Binary indicator for history of sitagliptin}
  \item{salmon.oil}{Binary indicator for history of salmon oil}
  \item{olmesartan}{Binary indicator for history of olmesartan}
  \item{timolol}{Binary indicator for history of timolol}
  \item{nitrofurantoin}{Binary indicator for history of nitrofurantoin}
  \item{celecoxib}{Binary indicator for history of celecoxib}
  \item{glimepiride}{Binary indicator for history of glimepiride}
  \item{iohexol}{Binary indicator for history of iohexol}
  \item{clonidine}{Binary indicator for history of clonidine}
  \item{valacyclovir}{Binary indicator for history of valacyclovir}
  \item{ropinirole}{Binary indicator for history of ropinirole}
  \item{bupivacaine}{Binary indicator for history of bupivacaine}
  \item{benzonatate}{Binary indicator for history of benzonatate}
  \item{carbamazepine}{Binary indicator for history of carbamazepine}
  \item{meclizine}{Binary indicator for history of meclizine}
  \item{azelastine}{Binary indicator for history of azelastine}
  \item{diclofenac}{Binary indicator for history of diclofenac}
  \item{clobetasol}{Binary indicator for history of clobetasol}
  \item{ketorolac}{Binary indicator for history of ketorolac}
  \item{amphetamine}{Binary indicator for history of amphetamine}
  \item{budesonide}{Binary indicator for history of budesonide}
  \item{diazepam}{Binary indicator for history of diazepam}
  \item{repaglinide}{Binary indicator for history of repaglinide}
  \item{omega.3.acid.ethyl.esters..USP.}{Binary indicator for history of omega-3-acid ethyl esters (USP)}
  \item{methadone}{Binary indicator for history of methadone}
  \item{clindamycin}{Binary indicator for history of clindamycin}
  \item{rivaroxaban}{Binary indicator for history of rivaroxaban}
  \item{vitamin.E}{Binary indicator for history of vitamin E}
  \item{valproate}{Binary indicator for history of valproate}
  \item{tiotropium}{Binary indicator for history of tiotropium}
  \item{temazepam}{Binary indicator for history of temazepam}
  \item{chlorhexidine}{Binary indicator for history of chlorhexidine}
  \item{hydroxychloroquine}{Binary indicator for history of hydroxychloroquine}
  \item{nystatin}{Binary indicator for history of nystatin}
  \item{olanzapine}{Binary indicator for history of olanzapine}
  \item{nicotine}{Binary indicator for history of nicotine}
  \item{mometasone}{Binary indicator for history of mometasone}
  \item{prednisolone}{Binary indicator for history of prednisolone}
  \item{estrogens..conjugated..USP.}{Binary indicator for history of conjugated estrogens (USP)}
  \item{mupirocin}{Binary indicator for history of mupirocin}
  \item{loratadine}{Binary indicator for history of loratadine}
  \item{fexofenadine}{Binary indicator for history of fexofenadine}
  \item{solifenacin}{Binary indicator for history of solifenacin}
  \item{irbesartan}{Binary indicator for history of irbesartan}
  \item{ephedrine}{Binary indicator for history of ephedrine}
  \item{isosorbide}{Binary indicator for history of isosorbide}
  \item{gadoterate.meglumine}{Binary indicator for history of gadoterate meglumine}
  \item{ubidecarenone}{Binary indicator for history of ubidecarenone}
  \item{verapamil}{Binary indicator for history of verapamil}
  \item{promethazine}{Binary indicator for history of promethazine}
  \item{doxazosin}{Binary indicator for history of doxazosin}
  \item{fluconazole}{Binary indicator for history of fluconazole}
  \item{haloperidol}{Binary indicator for history of haloperidol}
  \item{rivastigmine}{Binary indicator for history of rivastigmine}
  \item{labetalol}{Binary indicator for history of labetalol}
  \item{eszopiclone}{Binary indicator for history of eszopiclone}
  \item{insulin.detemir}{Binary indicator for history of insulin detemir}
  \item{aripiprazole}{Binary indicator for history of aripiprazole}
  \item{vancomycin}{Binary indicator for history of vancomycin}
  \item{thiamine}{Binary indicator for history of thiamine}
  \item{epinephrine}{Binary indicator for history of epinephrine}
  \item{amitriptyline}{Binary indicator for history of amitriptyline}
  \item{tolterodine}{Binary indicator for history of tolterodine}
  \item{cefazolin}{Binary indicator for history of cefazolin}
  \item{lisdexamfetamine}{Binary indicator for history of lisdexamfetamine}
  \item{risperidone}{Binary indicator for history of risperidone}
  \item{mirabegron}{Binary indicator for history of mirabegron}
  \item{magnesium.oxide}{Binary indicator for history of magnesium oxide}
  \item{niacin}{Binary indicator for history of niacin}
  \item{pramipexole}{Binary indicator for history of pramipexole}
  \item{zoledronic.acid}{Binary indicator for history of zoledronic acid}
  \item{sex}{Binary sex}
  \item{age}{Age (continuous)}
  \item{race}{Race}
  \item{days}{Time from index to event (continuous)}
  \item{index_date}{Index date of diagnosis}
  \item{treatment}{Binary treatment indicator}
  \item{outcome_AD_value}{Outcome binary indicator: AD}
  \item{outcome_AD_time}{Outcome time: AD}
  \item{outcome_ADRD_value}{Outcome binary indicator: ADRD}
  \item{outcome_ADRD_time}{Outcome time: ADRD}
  \item{outcome_acute_conjunctivitis_value}{Outcome binary indicator: acute conjunctivitis}
  \item{outcome_acute_conjunctivitis_time}{Outcome time: acute conjunctivitis}
  \item{outcome_acute_tonsillitis_value}{Outcome binary indicator: acute tonsillitis}
  \item{outcome_acute_tonsillitis_time}{Outcome time: acute tonsillitis}
  \item{outcome_adhesive_capsulitis_of_shoulder_value}{Outcome binary indicator: adhesive capsulitis of shoulder}
  \item{outcome_adhesive_capsulitis_of_shoulder_time}{Outcome time: adhesive capsulitis of shoulder}
  \item{outcome_allergic_rhinitis_value}{Outcome binary indicator: allergic rhinitis}
  \item{outcome_allergic_rhinitis_time}{Outcome time: allergic rhinitis}
  \item{outcome_blepharitis_value}{Outcome binary indicator: blepharitis}
  \item{outcome_blepharitis_time}{Outcome time: blepharitis}
  \item{outcome_carpal_tunnel_syndrome_value}{Outcome binary indicator: carpal tunnel syndrome}
  \item{outcome_carpal_tunnel_syndrome_time}{Outcome time: carpal tunnel syndrome}
  \item{outcome_chalazion_value}{Outcome binary indicator: chalazion}
  \item{outcome_chalazion_time}{Outcome time: chalazion}
  \item{outcome_contact_dermatitis_value}{Outcome binary indicator: contact dermatitis}
  \item{outcome_contact_dermatitis_time}{Outcome time: contact dermatitis}
  \item{outcome_dental_caries_value}{Outcome binary indicator: dental caries}
  \item{outcome_dental_caries_time}{Outcome time: dental caries}
  \item{outcome_deviated_nasal_septum_value}{Outcome binary indicator: deviated nasal septum}
  \item{outcome_deviated_nasal_septum_time}{Outcome time: deviated nasal septum}
  \item{outcome_foreign_body_in_ear_value}{Outcome binary indicator: foreign body in ear}
  \item{outcome_foreign_body_in_ear_time}{Outcome time: foreign body in ear}
  \item{outcome_gout_value}{Outcome binary indicator: gout}
  \item{outcome_gout_time}{Outcome time: gout}
  \item{outcome_hemorrhoids_value}{Outcome binary indicator: hemorrhoids}
  \item{outcome_hemorrhoids_time}{Outcome time: hemorrhoids}
  \item{outcome_impacted_cerumen_value}{Outcome binary indicator: impacted cerumen}
  \item{outcome_impacted_cerumen_time}{Outcome time: impacted cerumen}
  \item{outcome_influenza_value}{Outcome binary indicator: influenza}
  \item{outcome_influenza_time}{Outcome time: influenza}
  \item{outcome_ingrowing_nail_value}{Outcome binary indicator: ingrowing nail}
  \item{outcome_ingrowing_nail_time}{Outcome time: ingrowing nail}
  \item{outcome_low_back_pain_value}{Outcome binary indicator: low back pain}
  \item{outcome_low_back_pain_time}{Outcome time: low back pain}
  \item{outcome_menieres_disease_value}{Outcome binary indicator: menieres disease}
  \item{outcome_menieres_disease_time}{Outcome time: menieres disease}
  \item{outcome_osteoarthritis_of_knee_value}{Outcome binary indicator: osteoarthritis of knee}
  \item{outcome_osteoarthritis_of_knee_time}{Outcome time: osteoarthritis of knee}
  \item{outcome_osteoporosis_value}{Outcome binary indicator: osteoporosis}
  \item{outcome_osteoporosis_time}{Outcome time: osteoporosis}
  \item{outcome_foot_drop_value}{Outcome binary indicator: foot drop}
  \item{outcome_foot_drop_time}{Outcome time: foot drop}
  \item{outcome_hearing_problem_value}{Outcome binary indicator: hearing problem}
  \item{outcome_hearing_problem_time}{Outcome time: hearing problem}
  \item{outcome_intra_abdominal_and_pelvic_swelling_mass_and_lump_value}{Outcome binary indicator: intra abdominal and pelvic swelling mass and lump}
  \item{outcome_intra_abdominal_and_pelvic_swelling_mass_and_lump_time}{Outcome time: intra abdominal and pelvic swelling mass and lump}
  \item{outcome_irritability_and_anger_value}{Outcome binary indicator: irritability and anger}
  \item{outcome_irritability_and_anger_time}{Outcome time: irritability and anger}
  \item{outcome_wristdrop_value}{Outcome binary indicator: wristdrop}
  \item{outcome_wristdrop_time}{Outcome time: wristdrop}
  \item{site}{Study site identifier}
}
}
\usage{
LATTE_ADRD
}
\description{
A simulated ADRD data set for Lossless Oneshot Algorithm for Target Trial Emulation (LATTE)
}
\keyword{datasets}
