% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{long_covid}
\alias{long_covid}
\title{DisC2o simulated data}
\format{
A data frame with 900 rows and 53 variables:
\describe{
  \item{PASC_features}{number of Post Acute Sequelae of COVID (PASC, or long covid) features} 
  \item{covid_vaccination}{treatment of covid vaccination, 1=vaccinated}
  \item{site}{site id, 300 participants each for 'site1', 'site2', and 'site3'} 
  \item{X1}{a binary covariate}
  \item{X2}{a binary covariate}
  \item{X3}{a continuous covariate}  
  \item{X4}{a continuous covariate} 
  \item{X5}{a continuous covariate} 
}
}
\usage{
long_covid
}
\description{
A simulated long-covid data set for Distributed causal inference with covariates shift (DisC2o). This only contains 5 covariates and more noisy covariates can be added when running demo example.
}
\keyword{datasets}
