% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lung2}
\alias{lung2}
\title{Lung cancer survival time data}
\format{
A data frame with 228 rows and 5 variables:
\describe{
  \item{site}{simulated site id, 86 'site1', 83 'site2' and 59 'site3'}
  \item{time}{survival time in days}
  \item{status}{censoring status 0=censored, 1=dead}
  \item{age}{age in years}
  \item{sex}{1 for female and 0 for male}
}
}
\source{
\url{https://CRAN.R-project.org/package=survival}
}
\usage{
lung2
}
\description{
A data set modified from the lung data in survival package (see demo(ODAC)).
}
\keyword{datasets}
