% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistCox.R
\name{my_coxph}
\alias{my_coxph}
\title{Fit Cox PH model}
\usage{
my_coxph(data, beta_ini, hasTies = NULL)
}
\arguments{
\item{data}{List with component time (time-to-event), event (event indicator) and X (covariates).}

\item{beta_ini}{Vector of coef for initializing \code{optim()}, if not specified, use 0.}

\item{hasTies}{Logical - is the time to event tied? If yes, use Efron's approximation for tie correction in the partial likelihood.}
}
\value{
An \code{optim()} output.
}
\description{
fit Cox reg to time-to-event data. Expected to be the same as coxph() in package 'survival'. 
              Optimized by optim(), with log-L and gradient written in Rcpp.
}
\examples{
# see the example in DistCox()
}
\author{
Chongliang Luo, Rui Duan, Yong Chen
}
