% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmm.R
\name{myglmmPQL}
\alias{myglmmPQL}
\title{A flexible version of MASS::glmmPQL}
\usage{
myglmmPQL(formula.glm, formula, offset=NULL, family, data, fixef.init = NULL, 
                 weights=NULL, REML=T, niter=10, verbose=T)
}
\arguments{
\item{formula.glm}{formula used to fit \code{glm} for initial fixed effects}

\item{formula}{formula used to fit iterative \code{lmer} in PQL algorithm}

\item{offset}{\code{glm} offset term}

\item{family}{\code{glm} family}

\item{data}{\code{glm} data}

\item{fixef.init}{initial fixed effects estimates, set to zeros if NULL}

\item{weights}{\code{glm} weights}

\item{REML}{\code{lmer} logical scalar - Should the estimates be chosen to optimize the REML criterion (as opposed to the log-likelihood)?}

\item{niter}{\code{glmmPQL} maximum number of iterations.}

\item{verbose}{\code{glmmPQL} logical: print out record of iterations?}
}
\value{
An object wiht the same format as \code{lmer}.
}
\description{
A flexible version of MASS::glmmPQL
}
\details{
Use lme4::lmer instead of nlme::varFixed in PQL iteration to allow REML
}
\keyword{internal}
