% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODACAT.R
\name{ODACAT.initialize}
\alias{ODACAT.initialize}
\title{ODACAT initialize}
\usage{
ODACAT.initialize(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
init
}
\description{
ODACAT initialize
}
\keyword{internal}
