% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ODACAT_nominal}
\alias{ODACAT_nominal}
\title{ODACAT simulated data}
\format{
A data frame with 300 rows and 5 variables:
\describe{
  \item{id.site}{site id, 102 'site1', 100 'site2', 98 'site3'}
  \item{outcome}{3-category outcome, possible values are 1,2,3. Category 3 will be used as reference}
  \item{X1}{the first covariate, continuous}
  \item{X2}{the second covariate, binary}
  \item{X3}{the third covariate, binary}
}
}
\usage{
ODACAT_nominal
}
\description{
A simulated data set for ODACAT demonstration
}
\keyword{datasets}
