% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODACATH.R
\name{ODACATH.derive}
\alias{ODACATH.derive}
\title{ODACATH derivatives}
\usage{
ODACATH.derive(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
list(site=config$site_id, site_size = n, S_site=S_site, eta=eta_mat[site,])
}
\description{
ODACATH derivatives
}
\keyword{internal}
