% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODACATH.R
\name{ODACATH.estimate}
\alias{ODACATH.estimate}
\title{PDA surrogate estimation}
\usage{
ODACATH.estimate(ipdata,control,config)
}
\arguments{
\item{ipdata}{local data in data frame}

\item{control}{PDA control}

\item{config}{cloud configuration}
}
\value{
list(btilde=betanew, btilde.se=beta_SE,eta_mat=eta_mat,eta_mat_theta=NULL,site=config$site_id, site_size=n_site)
}
\description{
PDA surrogate estimation
}
\details{
step-3: construct and solve surrogate efficient score at the master/lead site
}
\keyword{internal}
