% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODACATH.R
\name{ODACATH.initialize}
\alias{ODACATH.initialize}
\title{ODACATH initialize}
\usage{
ODACATH.initialize(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
init
}
\description{
ODACATH initialize
}
\keyword{internal}
