% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODACH_CC.R
\name{ODACH_CC.derive}
\alias{ODACH_CC.derive}
\title{Generate pda derivatives}
\usage{
ODACH_CC.derive(ipdata, control, config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
list(bbar=bbar, site=control$mysite, site_size = nrow(ipdata), logL_D1=logL_D1, logL_D2=logL_D2)
}
\description{
Generate pda derivatives
}
\details{
Calculate and broadcast 1st and 2nd order derivative at initial bbar
}
\keyword{internal}
