% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODACH_CC.R
\name{ODACH_CC.estimate}
\alias{ODACH_CC.estimate}
\title{PDA surrogate estimation}
\usage{
ODACH_CC.estimate(ipdata, control, config)
}
\arguments{
\item{ipdata}{local data in data frame}

\item{control}{pda control}

\item{config}{cloud config}
}
\value{
list(btilde = sol$par, Htilde = sol$hessian, site=control$mysite, site_size=nrow(ipdata))
}
\description{
PDA surrogate estimation
}
\details{
step-4: construct and solve surrogate logL at the master/lead site
}
\keyword{internal}
