% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODACH_CC.R
\name{ODACH_CC.initialize}
\alias{ODACH_CC.initialize}
\title{ODACH_CC initialize}
\usage{
ODACH_CC.initialize(ipdata, control, config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
list(bhat_i = fit_i$coef, Vhat_i = summary(fit_i)$coef[,2]^2, site=control$mysite, site_size= nrow(ipdata))
}
\description{
ODACH_CC initialize
}
\references{
Chongliang Luo, et al. "ODACH: A One-shot Distributed Algorithm for Cox model with Heterogeneous Multi-center Data".
              medRxiv, 2021, https://doi.org/10.1101/2021.04.18.21255694
}
\keyword{internal}
