% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{odach_cc}
\alias{odach_cc}
\title{ODACH_CC simulated data}
\format{
A data frame with 413 rows and 8 variables:
\describe{
  \item{site}{site id, 187 'site1', 133 'site2', 93 'site3'. The full_cohort_size are 800, 600 and 400 respectively} 
  \item{subcohort}{1=subcohort, e.g. uniformly subsampled from full_cohort_size, 0=case} 
  \item{time}{survival time}
  \item{status}{censoring status 0=censored, 1=dead} 
  \item{X1}{the first covariate, continuous}
  \item{X2}{the second covariate, continuous} 
  \item{Category}{the third covariate, categorical} 
  \item{Group}{the fourth covariate, categorical} 
}
}
\usage{
odach_cc
}
\description{
A simulated data set for ODACH with case-cohort design demonstration
}
\keyword{datasets}
