\name{ODACO-package}
\alias{ODACO-package}
\alias{ODACO}
\docType{package}
\title{
\packageTitle{ODACO}
}
\description{
\packageDescription{ODACO}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ODACO}
\packageIndices{ODACO}
Fit a Communication-Efficient Distributed Cox Model by optimizing a surrogate log-likelihood using the partial likelihood of the local machine and the gradients and hessian of remote machines.}
\author{
\packageAuthor{ODACO}

Maintainer: \packageMaintainer{ODACO}
}
\references{
Jordan, Michael I., Jason D. Lee, and Yun Yang. "Communication-efficient distributed statistical inference." JASA (2018).
}
\keyword{ package }
\seealso{
}
\examples{
}
