% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODACT.R
\name{ODACT.derive}
\alias{ODACT.derive}
\title{Generate pda ODACT derivatives}
\usage{
ODACT.derive(ipdata, control, config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
list(b_meta=b_meta, site=control$mysite, site_size = nrow(ipdata), logL_D1=logL_D1, logL_D2=logL_D2)
}
\description{
Generate pda ODACT derivatives
}
\details{
Calculate and broadcast 1st and 2nd order derivative at initial bbar for ODACT
}
\keyword{internal}
