% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODACT.R
\name{ODACT.estimate}
\alias{ODACT.estimate}
\title{PDA ODACT surrogate estimation}
\usage{
ODACT.estimate(ipdata, control, config)
}
\arguments{
\item{ipdata}{local data in data frame}

\item{control}{pda control}

\item{config}{cloud config}
}
\value{
list(btilde = sol$par, Htilde = sol$hessian, site=control$mysite, site_size=nrow(ipdata))
}
\description{
PDA ODACT surrogate estimation
}
\details{
step-4: construct and solve surrogate logL at the master/lead site
}
\keyword{internal}
