% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODACT.R
\name{ODACT.initialize}
\alias{ODACT.initialize}
\title{ODACT initialize}
\usage{
ODACT.initialize(ipdata, control, config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
list(bhat_i, Vhat_i, site, site_size)
}
\description{
ODACT initialize
}
\references{
Liang CJ, Luo C, Kranzler HR, Bian J, Chen Y. Communication-efficient federated learning of temporal effects on opioid use disorder with data from distributed research networks. J Am Med Inform Assoc. 2025 Apr 1;32(4):656-664. doi: 10.1093/jamia/ocae313. PMID: 39864407; PMCID: PMC12005629.
}
\keyword{internal}
