% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAH.R
\name{ODAH.derive}
\alias{ODAH.derive}
\title{ODAH derivatives}
\usage{
ODAH.derive(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
derivatives list(site = config$site_id, site_size = nrow(ipdata), 
                          logL_D1_zero = logL_D1_zero, logL_D1_count = logL_D1_count,
                          logL_D2_zero = logL_D2_zero, logL_D2_count = logL_D2_count)
}
\description{
ODAH derivatives
}
\keyword{internal}
