% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAH.R
\name{ODAH.estimate}
\alias{ODAH.estimate}
\title{PDA surrogate estimation}
\usage{
ODAH.estimate(ipdata,control,config)
}
\arguments{
\item{ipdata}{local data in  a list(ipdata, X_count, X_zero)}

\item{control}{PDA control}

\item{config}{cloud configuration}
}
\value{
list(btilde = sol$par, Htilde = sol$hessian, site=control$mysite, site_size=nrow(ipdata))
}
\description{
PDA surrogate estimation
}
\details{
construct and solve surrogate logL at the master/lead site
}
\keyword{internal}
