% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAH.R
\name{ODAH.initialize}
\alias{ODAH.initialize}
\title{ODAH initialize}
\usage{
ODAH.initialize(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
init
}
\description{
ODAH initialize
}
\references{
TBD
}
\keyword{internal}
