% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAL.R
\name{ODAL.derive}
\alias{ODAL.derive}
\title{ODAL derivatives}
\usage{
ODAL.derive <- function(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
list(T_all=T_all, b_meta=b_meta, site=control$mysite, site_size = nrow(mydata), U=U, W=W, Z=Z, logL_D1=logL_D1, logL_D2=logL_D2)
}
\description{
ODAL derivatives
}
