% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAL.R
\name{ODAL.estimate}
\alias{ODAL.estimate}
\title{PDA surrogate estimation}
\usage{
ODAL.estimate(ipdata,control,config)
}
\arguments{
\item{ipdata}{local data in data frame}

\item{control}{PDA control}

\item{config}{cloud configuration}
}
\value{
list(btilde = sol$par, Htilde = sol$hessian, site=control$mysite, site_size=nrow(ipdata))
}
\description{
PDA surrogate estimation
}
\details{
step-3: construct and solve surrogate logL at the master/lead site
}
\keyword{internal}
