% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAL.R
\name{ODAL.initialize}
\alias{ODAL.initialize}
\title{ODAL initialize}
\usage{
ODAL.initialize <- function(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
init
}
\description{
ODAL initialize
}
