% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAL.R
\name{ODAL.synthesize}
\alias{ODAL.synthesize}
\title{PDA synthesize surrogate estimates from all sites, optional}
\usage{
ODAL.synthesize(ipdata,control,config)
}
\arguments{
\item{ipdata}{local data in data frame}

\item{control}{pda control}

\item{config}{pda cloud configuration}
}
\value{
list(btilde=btilde,  Vtilde=Vtilde)
}
\description{
PDA synthesize surrogate estimates from all sites, optional
}
\details{
Optional step-4: synthesize all the surrogate est btilde_i from each site, if step-3 from all sites is broadcasted
}
\keyword{internal}
