% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAP.R
\name{ODAP.derive}
\alias{ODAP.derive}
\title{ODAP derivatives}
\usage{
ODAP.derive(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
derivatives list(site = config$site_id, site_size = nrow(ipdata), logL_D1 = logL_D1, logL_D2 = logL_D2) 
                     or 
                     list(site = config$site_id, site_size = nrow(ipdata), 
                          logL_D1_zero = logL_D1_zero, logL_D1_count = logL_D1_count,
                          logL_D2_zero = logL_D2_zero, logL_D2_count = logL_D2_count)
                     when family='hurdle'.
}
\description{
ODAP derivatives
}
\keyword{internal}
