% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAP.R
\name{ODAP.estimate}
\alias{ODAP.estimate}
\title{PDA surrogate estimation}
\usage{
ODAP.estimate(ipdata,control,config)
}
\arguments{
\item{ipdata}{local data in data frame (generated in \code{pda}, is a list(ipdata, X_count, X_zero) if family='hurdle')}

\item{control}{PDA control}

\item{config}{cloud configuration}
}
\value{
list(btilde = sol$par, Htilde = sol$hessian, site=control$mysite, site_size=nrow(ipdata))
}
\description{
PDA surrogate estimation
}
\details{
construct and solve surrogate logL at the master/lead site
}
\keyword{internal}
