% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAP.R
\name{ODAP.initialize}
\alias{ODAP.initialize}
\title{ODAP initialize}
\usage{
ODAP.initialize(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
init
}
\description{
ODAP initialize
}
\references{
TBD
}
\keyword{internal}
