% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAPB.R
\name{ODAPB.derive}
\alias{ODAPB.derive}
\title{ODAPB derivatives}
\usage{
ODAPB.derive(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
derivatives list(site = config$site_id, site_size = nrow(ipdata), logL_D1 = logL_D1, logL_D2 = logL_D2)
}
\description{
ODAPB derivatives
}
\keyword{internal}
