% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAPB.R
\name{ODAPB.initialize}
\alias{ODAPB.initialize}
\title{ODAPB initialize}
\usage{
ODAPB.initialize(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
init
}
\description{
ODAPB initialize
}
\references{
TBD
}
\keyword{internal}
