% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ordinal_data}
\alias{ordinal_data}
\title{ODACATH simulated data with ordinal outcome}
\format{
A list containing the following elements:
\describe{
  \item{id.site}{site id, 100 'site1', 100 'site2', 100 'site3'}
  \item{y}{ordinal outcome taking values 1,2,3}
  \item{X1_cont}{a continuous covariate} 
  \item{X2_bin}{a binary covariate} 
}
}
\usage{
ordinal_data
}
\description{
A simulated data set for ODACATH demonstration
}
\keyword{datasets}
