\name{pda-internal}
\alias{pda-internal}
\alias{all.covariates}
\alias{arc.arrow}
\alias{arrow}
\alias{budworm}
\alias{dotip}
\alias{expt}
\alias{nested.means}
\alias{path3}
\alias{pdiff.ravel}
\alias{pletter.list}
\alias{pletter.ravel}
\alias{power.curve}
\alias{projref}
\alias{projwt}
\alias{splus.data}
\alias{subsamp}
\alias{terms.lme}
\alias{terms.lmer}
\title{Internal pda functions.}
\description{
  These are generally not to be called by the user.
}
\usage{
all.covariates()
arc.arrow()
arrow()
data(budworm)
dotip()
expt()
fact()
nested.means()
path3()
pdiff.ravel()
perms()
pletter.list()
pletter.ravel()
power.curve()
projref()
projwt()
splus.data()
subsamp()
\method{terms}{lme}()
\method{terms}{lmer}()
}
\author{ Brian S. Yandell, yandell@stat.wisc.edu }
\keyword{internal}
