\name{pda-package}
\alias{pda-package}
\alias{pda}
\docType{package}
\title{
\packageTitle{pda}
}
\description{
\packageDescription{pda}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pda}
\packageIndices{pda}
Fit a Communication-Efficient Distributed Cox Model by optimizing a surrogate log-likelihood using the partial likelihood of the local machine and the gradients and hessian of remote machines.}
\author{
\packageAuthor{pda}

Maintainer: \packageMaintainer{pda}
}
\references{
Jordan, Michael I., Jason D. Lee, and Yun Yang. "Communication-efficient distributed statistical inference." JASA (2018).
}
\keyword{ package }
\seealso{
}
\examples{
}
