% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.R
\name{pda}
\alias{pda}
\title{PDA: Privacy-preserving Distributed Algorithm}
\usage{
pda(data = ipdata, mysite = NULL)
}
\arguments{
\item{ipdata}{Local IPD data in data frame, should include at least one column for the outcome and one column for the covariates}

\item{site_id}{Character site name}

\item{control}{pda control data}
}
\description{
Fit Privacy-preserving Distributed Algorithms for linear, logistic, 
               Poisson and Cox PH regression with possible heterogeneous data across sites.
}
\examples{
require(survival)
require(data.table)
require(pda)
data(lung)
#create a number of sites, split the lung data amongst them
sites = c('site1', 'site2', 'site3')
set.seed(42)
lung2<-lung[,2:5]
lung2$sex <- lung2$sex-1
lung2$status <- ifelse(lung2$status == 2, 1, 0)
lung_split<-split(lung2, sample(1:length(sites), nrow(lung), replace=T))
######################### setup  ODAL #############################
control <- list(project_name = 'Lung cancer study',
       step = 'initialize',    # current step, updated by lead
       sites = sites,
       heterogeneity = FALSE,
       model = 'ODAL',
       outcome = "status",
       variables = c('age', 'sex'),
       optim_maxit=100,
       lead_site = sites[1],
       upload_date = as.character(Sys.time()),
       heterogeneity = FALSE)
## RUN BY LEAD ONLY 
pda(site_id='site1',control=control)
#run pda until step is empty
while (is.character(control$step)) {
 print(paste("step:",control$step))
 #cycle through sites
 for(i in 1:length(sites)) {
   control<-pda(ipdata=lung_split[[i]],site_id=sites[i])
 }
}

}
\references{
Jordan, Michael I., Jason D. Lee, and Yun Yang. "Communication-efficient distributed statistical inference." JASA (2018).
}
\author{
Chongliang Luo, Steven Vitale, Jiayi Tong, Rui Duan, Yong Chen
}
