% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.R
\name{pdaCatalog}
\alias{pdaCatalog}
\title{use this function to guide end-users step-by-step to identify best pda models for their tasks, and set up control.}
\usage{
pdaCatalog(task=c('Regression', 'Survival', 'Trial_emulation', 
'Causal_inference', 'Design_analysis', 'Clustering'), 
write_json_file_path=getwd(), optim_maxit,optim_method,init_method)
}
\arguments{
\item{task}{user-specified task, c('Regression', 'Survival', 'Trial_emulation', 'Causal_inference', 'Design_analysis', 'Clustering'). If no specify, display all models}

\item{write_json_file_path}{directory path to write the control file to}

\item{optim_maxit}{option in the control file for the optimization in pda, default 100}

\item{optim_method}{option in the control file for the optimization in pda, default "BFGS"}

\item{init_method}{option in the control file for calculating the initial estimate in pda, default "meta"}
}
\value{
pda control
}
\description{
use this function to guide end-users step-by-step to identify best pda models for their tasks, and set up control.
}
\seealso{
\code{pda}
}
