\name{PDAexamples}
\alias{PDAexamples}
\title{PDA examples.}
\non_function{}
\description{
Examples from the book Practical Data Analysis with Designed
Experiments by Brian S. Yandell (1997 Chapman \& Hall/CRC Press).
}
\examples{
## Book examples with R code (book parts after #)
example(Bacteria) # C,H
example(BactRoom) # C
example(Berry)    # H
example(BrandX)   #
example(Brassica) # H
example(Budworm)  #
example(Cloning)  # B,E
example(Diet)     # B,F
example(Drink)    # I
example(Drying)   #
example(Feed)     # F
example(Forage)   # H
example(Growth)   # D,E
example(GrowthCount) #
example(Hardy)    # D,F,H
example(Oocyte)   # D
example(Random)   # G
example(Rantwo)   # 
example(Running)  # C
example(Tomato)   # B,H
example(Tree)     # I
example(Tukey)    # C
example(Wasp)     # F
example(WheatPDA) # G

## Other book examples that currently have no R code
#example(Alfalfa) # H
#example(BST)     # D
#example(Biotron) # C
#example(Budding) # C
#example(Chew)    # I
#example(Company) # C
#example(Damage)  # F
#example(Design)  # A
#example(Genss)   # E
#example(Grain)   # H
#example(Iris)    # F
#example(Iron)    # F
#example(Leprosy) # F
#example(Paddock) # I
#example(Potato)  # H
#example(Prefer)  # C
#example(Product) # D
#example(Rumen)   # I
#example(Season)  # I
#example(Select)  # G
#example(Strip)   # H
#example(Variance)# E
}
\author{ Brian S. Yandell, yandell@stat.wisc.edu }
\keyword{datasets}
