% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.R
\name{pdaGet}
\alias{pdaGet}
\title{Function to download json and return as object}
\usage{
pdaGet(name)
}
\arguments{
\item{name}{of file}
}
\value{

}
\description{
Function to download json and return as object
}
\author{
Chongliang Luo, Steven Vitale
}
