% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.R
\name{pdaList}
\alias{pdaList}
\title{Function to list available objects}
\usage{
pdaList(config)
}
\arguments{
\item{config}{a list of variables for cloud configuration}
}
\value{
list of files
}
\description{
Function to list available objects
}
