% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.R
\name{pdaPut}
\alias{pdaPut}
\title{Function to upload object to cloud as json}
\usage{
pdaPut(obj,name,cloud_config)
}
\arguments{
\item{obj}{R object to encode as json and uploaded to cloud}

\item{name}{of file}

\item{cloud_config}{a list of variables for cloud configuration}
}
\value{

}
\description{
Function to upload object to cloud as json
}
\author{
Chongliang Luo, Steven Vitale
}
