% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latte-misc.R
\name{run_pooled_analysis}
\alias{run_pooled_analysis}
\title{Function to perform all data processing and pooled stratified analysis}
\usage{
run_pooled_analysis(data, outcome_id, outcome_time, sites)
}
\arguments{
\item{data}{The combined LATTE_ADRD data.}

\item{outcome_id}{The name of the primary outcome column.}

\item{outcome_time}{The name of the outcome time column (for Poisson).}

\item{sites}{A vector of site identifiers.}
}
\value{
A list containing the results of the standard logistic and Poisson pooled analysis.
}
\description{
Function to perform all data processing and pooled stratified analysis
}
