\name{splitplot}
\alias{splitplot}
\title{~function to do ???}
\usage{
splitplot(response, nest.factors, covars, weights, data)
}
\description{}
\arguments{
    \item{response}{~Describe response here}
    \item{nest.factors}{~Describe nest.factors here}
    \item{covars}{~Describe covars here}
    \item{weights}{~Describe weights here}
    \item{data}{~Describe data here}}
\value{~Describe the value returned}
\examples{
}
\keyword{htest}
% Converted by  version .
