\name{std.dev}
\alias{std.dev}
\title{standard deviation of object}
\usage{std.dev(object)}
\description{}
\arguments{
    \item{object}{vector or object from linear model fit}}
\value{standard deviation}
\examples{
\dontrun{
std.dev(x)
std.dev(lm(y~a+b,data))
}
}
\keyword{htest}
% Converted by  version .
