\name{unfactor}
\alias{unfactor}
\title{unclass factor (but keep numeric level)}
\usage{unfactor(x.factor, class="factor")}
\description{}
\arguments{
    \item{x.factor}{element of class "factor" or "ordered"}
    \item{class}{usually "factor" or "ordered"}}
\value{vector of elements of x.factor}
\seealso{\code{\link[base]{unclass}}, \code{\link[base]{interaction.plot}}.}
\examples{
\dontrun{
unfactor(a)
}
}
% Converted by  version .
\keyword{manip}
