% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.R
\name{getCloudConfig}
\alias{getCloudConfig}
\title{gather cloud settings into one list}
\usage{
getCloudConfig()

getCloudConfig()
}
\arguments{
\item{site_id}{site identifier}

\item{dir}{shared directory path if using flat files}

\item{uri}{web uri if using web service}

\item{secret}{web token if using web service}
}
\description{
gather cloud settings into one list
}
\author{
Chongliang Luo, Steven Vitale
}
