% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.R
\name{getCloudConfig}
\alias{getCloudConfig}
\title{gather cloud settings into a list}
\usage{
getCloudConfig(site_id,dir,uri,secret)
}
\arguments{
\item{site_id}{site identifier}

\item{dir}{shared directory path if flat files}

\item{uri}{web uri if web service}

\item{secret}{web token if web service}
}
\value{
A list of cloud parameters: site_id, secret and uri
}
\description{
gather cloud settings into a list
}
\seealso{
\code{pda}
}
