% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAC.R
\name{ODAC.derive}
\alias{ODAC.derive}
\title{Generate pda derivatives}
\usage{
ODAC.derive <- function(ipdata, control, config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
list(T_all=T_all, b_meta=b_meta, site=control$mysite, site_size = nrow(ipdata), U=U, W=W, Z=Z, logL_D1=logL_D1, logL_D2=logL_D2)
}
\description{
Generate pda derivatives
}
