% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAC.R
\name{ODAC.initialize}
\alias{ODAC.initialize}
\title{ODAC initialize}
\usage{
ODAC.initialize <- function(ipdata, control, config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
list(T_i = T_i, bhat_i = fit_i$coef, Vhat_i = summary(fit_i)$coef[,2]^2, site=control$mysite, site_size= nrow(ipdata))
}
\description{
ODAC initialize
}
