% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ODACAT_nominal}
\alias{ODACAT_nominal}
\title{ODACAT simulated data with nominal outcome}
\format{
A list containing the following elements:
\describe{
  \item{id.site}{site id, 100 'site1', 100 'site2', 100 'site3'}
  \item{outcome}{nominal outcome taking values 1,2,3}
  \item{X1}{a continuous covariate} 
  \item{X2}{a binary covariate} 
  \item{X3}{a binary covariate}
}
}
\usage{
ODACAT_nominal
}
\description{
A simulated data set for ODACAT demonstration
}
\keyword{datasets}
