% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAL.R
\name{ODAL.derive}
\alias{ODAL.derive}
\title{ODAL derivatives}
\usage{
ODAL.derive(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
list(site=config$site_id, site_size = nrow(ipdata), logL_D1=logL_D1, logL_D2=logL_D2)
}
\description{
ODAL derivatives
}
\keyword{internal}
