% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAL.R
\name{ODAL.initialize}
\alias{ODAL.initialize}
\title{ODAL initialize}
\usage{
ODAL.initialize(ipdata,control,config)
}
\arguments{
\item{ipdata}{individual participant data}

\item{control}{pda control data}

\item{config}{local site configuration}
}
\value{
init
}
\description{
ODAL initialize
}
\references{
Rui Duan, et al. "Learning from electronic health records across multiple sites: A communication-efficient and privacy-preserving distributed algorithm". 
               Journal of the American Medical Informatics Association, 2020, https://doi.org/10.1093/jamia/ocz199
}
\keyword{internal}
