% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.R
\name{pdaGet}
\alias{pdaGet}
\title{Function to download json and return as object}
\usage{
pdaGet(name,config)
}
\arguments{
\item{name}{of file}

\item{config}{cloud configuration}
}
\value{
data from json file
}
\description{
Function to download json and return as object
}
