% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.R
\name{pdaSync}
\alias{pdaSync}
\title{update the PDA control, used by the master site}
\usage{
pdaSync <- function(config,site_id)
}
\arguments{
\item{control_update}{Logical, update the PDA control?}
}
\description{
update the PDA control, used by the master site
}
