% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.R
\name{pdaSync}
\alias{pdaSync}
\title{pda control synchronize}
\usage{
pdaSync(config)
}
\arguments{
\item{config}{cloud configuration}
}
\value{
control
}
\description{
update pda control if ready (run by lead)
}
\seealso{
\code{pda}
}
